+++ packages/pudu/v0.1/src/Pudu/Eval/Install.hs
@@ -137,6 +137,7 @@ installBuiltinConstructors = do
   bind "swissTableSize" (BuiltinValue SwissTableSizeBuiltin)
   bind "csvRecords" (BuiltinValue CsvRecordsBuiltin)
   bind "jsonDecode" (BuiltinValue JsonDecodeBuiltin)
+  bind "jsonEncode" (BuiltinValue JsonEncodeBuiltin)
   bind "bufferReadI64" (BuiltinValue BufferReadI64Builtin)
   bind "bufferWriteI64" (BuiltinValue BufferWriteI64Builtin)
   bind "bufferReadF64" (BuiltinValue BufferReadF64Builtin)